Furkan_Work/ai_detection/frame_splitter/requirements.txt
opencv-python
kafka-python
redis
numpy
